fix error

+++ systems/gogoll_expert_system.py
@@ -154,8 +154,8 @@ class GogollExpertSystem(pl.LightningModule):
             y_seg_t_t = self.seg_t(target_img)
             y_seg_t_t_c = self.seg_t(cycled_target)
 
-            loss_seg_a = self.semseg_loss(y_seg_s, segmentation_img)
-            loss_seg_b = self.semseg_loss(y_seg_t, segmentation_img)
+            loss_seg_a = self.semseg_loss(y_seg_s, segmentation_img.long())
+            loss_seg_b = self.semseg_loss(y_seg_t, segmentation_img.long())
             loss_seg_c = self.semseg_loss(y_seg_t_t, y_seg_t_s.argmax(dim=1).long())
             loss_seg_d = self.semseg_loss(y_seg_t_t_c, y_seg_t_s.argmax(dim=1).long())
